# [MRG + 2] Performance improvement for RLE decoder