#### What does this implement/fix? Explain your changes.
Improves the performance of the RLE plane decoder by eliminating a list construction and an arithmetic operation.

MR_small_RLE.dcm is about 25% faster (2.8 s vs 3.7 s for n = 10000).
emri_small_RLE.dcm is about 35% faster (11.5 s vs 18.3 s for n = 10000).