@scaramallion 2017-08-16
Both python 2 and 3 are faster, by the same amount.