@glemaitre 2017-08-16
I was wondering if you could give the minimal example of you performance test.
I wanted to check another way but I am not sure which code need to be called.