@scaramallion 2017-08-16
Its pretty ugly, but sure:
```python
#!/usr/bin/env python
import os
import timeit

from pydicom.data import DATA_ROOT
from pydicom.encaps import decode_data_sequence, defragment_data
from pydicom.pixel_data_handlers.rle_handler import _rle_decode_frame

test_file = os.path.join(DATA_ROOT, 'test_files', 'MR_small_RLE.dcm')
ds = read_file(test_file, force=True)

def wrapper(func, *args, **kwargs):
    def wrapped():
        return func(*args, **kwargs)
    return wrapped

def comparison_function(...):
    # Insert code here

if 'NumberOfFrames' in ds and ds.NumberOfFrames > 1:
    pixel_data = decode_data_sequence(ds.PixelData)
    for frame in pixel_data:
        wrapped = wrapper(_rle_decode_frame, frame, ds.Rows, ds.Columns, ds.SamplesPerPixel, ds.BitsAllocated)
        print(timeit.timeit(wrapped, number=1000))
else:
    frame = defragment_data(ds.PixelData)
    wrapped = wrapper(_rle_decode_frame, frame, ds.Rows, ds.Columns, ds.SamplesPerPixel, ds.BitsAllocated)
        print(timeit.timeit(wrapped, number=1000))
```
Then I'd swap out `_decode_rle_frame` with the `comparison_function`. Like I said, ugly.

The final test was calling `_decode_rle_frame` and switching between the master and perf-* branches.