@glemaitre 2017-08-16
The size of the array is probably to small to show the benefit of what I was thinking.
I thought of allocating `result` once for all. I would assume that python underneath is extending the size of the array twice whenever there is not enough space left when calling `extend`.

Anyway the implementation would look like that:

```python
def _rle_decode_plane(data, plane_size):                                                                             
    """Decodes a single plane of RLE encoded data using the PackBits algorithm                                       
                                                                                                                     
    Parameters                                                                                                       
    ----------                                                                                                       
    data : bytes                                                                                                     
        The data to be decompressed.                                                                                 
                                                                                                                     
    plane_size : int                                                                                                 
        The size of the plane.                                                                                       
                                                                                                                     
    Returns                                                                                                          
    -------                                                                                                          
    result : bytearray                                                                                               
        The decompressed data.                                                                                       
                                                                                                                     
    """                                                                                                              
                                                                                                                     
    data = bytearray(data)                                                                                           
    result = bytearray(plane_size)                                                                                   
    pos = 0                                                                                                          
    pos_res = 0                                                                                                      
    len_data = len(data)                                                                                             
                                                                                                                     
    while pos < len_data:                                                                                            
        header_byte = data[pos]                                                                                      
        pos += 1                                                                                                     
                                                                                                                     
        if header_byte > 128:                                                                                        
            result[pos_res:pos_res + (257 - header_byte)] = data[pos:pos + 1] * (257 - header_byte)                  
            pos_res += (257 - header_byte)                                                                           
            pos += 1                                                                                                 
        else:                                                                                                        
            result[pos_res:pos_res + header_byte + 1] = data[pos:pos + header_byte + 1]                              
            pos_res += header_byte + 1                                                                               
            pos += header_byte + 1                                                                                   
                                                                                                                     
    return result
```