@scaramallion 2017-08-17
Yeah, I literally tried that exact same thing, but it ended up being slower than the default.